ref(replays): Rename files related to the replay list page
